fonts

Module font

Structs

CachedShapeData 🔒

FallbackKey

Font

FontBaseline

FontFamilyDescriptor

The font family parameters for font selection.

FontGroup

A FontGroup is a prioritised list of fonts for a given set of font styles. It is used by TextRun to decide which font to render a character with. If none of the fonts listed in the styles are suitable, a fallback font may be used.

FontGroupFamily 🔒

A FontGroupFamily is a single font family in a FontGroup. It corresponds to one of the families listed in the font-family CSS property. The corresponding font data is lazy-loaded, only if actually needed. A single FontGroupFamily can have multiple fonts, in the case that individual fonts only cover part of the Unicode range.

FontGroupFamilyTemplate 🔒

A FontGroupFamily can have multiple associated FontTemplates if it is a “composite face”, meaning that it is defined by multiple @font-face declarations which vary only by their unicode-range descriptors. In this case, font selection will select a single member that contains the necessary unicode character. Unicode ranges are specified by the FontGroupFamilyTemplate::template member.

FontMetrics

FontRef

NoUsableFamilyName 🔒

An error indicating that the name table contained no usable family names.

ShapeCacheEntry 🔒

An entry in the shape cache.

ShapingFlags

ShapingOptions

Various options that control text shaping.

VariationAxes 🔒

Enums

FontSearchScope

The scope within which we will look for a font.

Constants

AFRC 🔒

BASE 🔒

CALT 🔒

CBDT 🔒

CLIG 🔒

COLR 🔒

CWSH 🔒

DLIG 🔒

FRAC 🔒

FVAR 🔒

FWID 🔒

GPOS 🔒

GSUB 🔒

HIST 🔒

HLIG 🔒

ITAL 🔒

JP04 🔒

JP78 🔒

JP83 🔒

JP90 🔒

KERN 🔒

LAST_RESORT_GLYPH_ADVANCE

LIGA 🔒

LNUM 🔒

NALT 🔒

NAME 🔒

ONUM 🔒

ORDN 🔒

ORNM 🔒

PNUM 🔒

PWID 🔒

RUBY 🔒

SALT 🔒

SBIX 🔒

SLNT 🔒

SMPL 🔒

SUBS 🔒

SUPS 🔒

SWSH 🔒

TNUM 🔒

TRAD 🔒

ZERO 🔒

Traits

FontTableMethods 🔒

PlatformFontMethods

Functions

compute_variations 🔒

https://drafts.csswg.org/css-fonts-4/#apply-font-matching-variations

map_platform_values_to_style_values 🔒

Given a mapping array mapping and a value, map that value onto the value specified by the array. For instance, for FontConfig values of weights, we would map these onto the CSS [0..1000] range by creating an array as below. Values that fall between two mapped values, will be adjusted by the weighted mean.

Type Aliases

FractionalPixel 🔒