fonts

Module font_context

Structs

FontCacheKey 🔒

FontContext

The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the system font service where required.

FontFaceRuleInitiator 🔒

FontFaceRuleWithOrigin 🔒

FontGroupCacheKey 🔒

FontParameters 🔒

KnownFontFaceRule 🔒

KnownFontFaceRules 🔒

RemoteWebFontDownloader 🔒

WebFontDocumentContext

Document-specific data required to fetch a web font.

WebFontDownloadState 🔒

Tracks the progress of loading a single @font-face rule by trying all specified sources in order.

Enums

DownloaderResponseResult 🔒

WebFontLoadInitiator 🔒

Statics

SMALL_CAPS_SCALE_FACTOR 🔒

Traits

CspViolationHandler

A callback that will be invoked on the Fetch thread if a web font download results in CSP violations. This handler will be cloned each time a new web font download is initiated.

FontContextWebFontMethods

NetworkTimingHandler

A callback that will be invoked on the Fetch thread when a web font download succeeds, providing timing information about the request.

Functions

font_face_rules_conflict 🔒

Returns true if the two @font-face rules cannot both apply at the same time.

Type Aliases

FontGroupRef

ScriptWebFontLoadFinishedCallback 🔒