Module fonts::font_template

Structs

FontTemplate

This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

FontTemplateDescriptor

Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

Traits

FontMatchDistanceMethod 🔒

A trait for implementing the CSS font matching algorithm against various font features. See https://drafts.csswg.org/css-fonts/#font-matching-algorithm.

FontTemplateRefMethods

IsOblique 🔒

Type Aliases

FontTemplateRef

A reference to a FontTemplate with shared ownership and mutability.