fonts::platform::freetype::font

Struct PlatformFont

pub struct PlatformFont {
    face: ReentrantMutex<FreeTypeFace>,
    requested_face_size: Au,
    actual_face_size: Au,
    variations: Vec<FontVariation>,
    synthetic_bold: bool,
    table_provider_data: FreeTypeFaceTableProviderData,
}

Fields
face: ReentrantMutex<FreeTypeFace>requested_face_size: Auactual_face_size: Auvariations: Vec<FontVariation>synthetic_bold: booltable_provider_data: FreeTypeFaceTableProviderData

A member that allows using skrifa to read values from this font.

Implementations

impl PlatformFont

fn unscalable_font_metrics_scale(&self) -> f64

Find the scale to use for metrics of unscalable fonts. Unscalable fonts, those using bitmap glyphs, are scaled after glyph rasterization. In order for metrics to match the final scaled font, we need to scale them based on the final size and the actual font size.