fonts

Module shapers

Modules

harfbuzz 🔒

Structs

ShapedGlyph 🔒

Constants

HIGHEST_DEFINED_CHARACTER_VARIANT_INDEX 🔒

The highest acceptable value for character-variant values in font-variant-alternates.

HIGHEST_DEFINED_STYLESET_INDEX 🔒

The highest acceptable value for styleset values in font-variant-alternates.

Traits

GlyphShapingResult 🔒

Holds the results of shaping. Abstracts over HarfBuzz and HarfRust which return data in very similar form but with different types

Functions

compute_used_font_features 🔒

Determine which OpenType features are applied for the font.

unicode_script_to_iso15924_tag 🔒

Utility function to convert a unicode_script::Script enum into the corresponding c_uint tag that harfbuzz uses to represent unicode scipts.