fonts::shapers

Module harfbuzz

Structs

FontFuncs 🔒

Callbacks from Harfbuzz when font map and glyph advance lookup needed.

HarfbuzzGlyphShapingResult 🔒

ShapedGlyphIterator 🔒

Shaper 🔒

Constants

HB_OT_TAG_DEFAULT_LANGUAGE 🔒

HB_OT_TAG_DEFAULT_SCRIPT 🔒

Statics

HB_FONT_FUNCS 🔒

Functions

destroy_blob_func 🔒

font_table_func 🔒

Callback to get a font table out of a font.

glyph_func 🔒

glyph_h_advance_func 🔒