fonts::shapers

Trait GlyphShapingResult

pub(crate) trait GlyphShapingResult {
    // Required methods
    fn len(&self) -> usize;
    fn is_rtl(&self) -> bool;
    fn iter(&self) -> impl Iterator<Item = ShapedGlyph>;
}

Holds the results of shaping. Abstracts over HarfBuzz and HarfRust which return data in very similar form but with different types

Required Methods

fn len(&self) -> usize

The number of shaped glyphs

fn is_rtl(&self) -> bool

Whether or not the result is right-to-left.

fn iter(&self) -> impl Iterator<Item = ShapedGlyph>

An iterator of the shaped glyphs of this data.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.