fonts::shapers

Trait HarfBuzzShapedGlyphData

trait HarfBuzzShapedGlyphData {
    // Required methods
    fn len(&self) -> usize;
    fn byte_offset_of_glyph(&self, i: usize) -> u32;
    fn entry_for_glyph(&self, i: usize, y_pos: &mut Au) -> ShapedGlyphEntry;
}

Holds the results of shaping. Abstracts over HarfBuzz and HarfRust which return data in very similar form but with different types

Required Methods

fn len(&self) -> usize

The number of shaped glyphs

fn byte_offset_of_glyph(&self, i: usize) -> u32

The byte offset of the shaped glyph in the souce text

fn entry_for_glyph(&self, i: usize, y_pos: &mut Au) -> ShapedGlyphEntry

Returns shaped glyph data for one glyph, and updates the y-position of the pen.