Module fonts::system_font_service

Structs

CSSFontFaceDescriptors

This data structure represents the various optional descriptors that can be applied to a @font-face rule in CSS. These are used to create a FontTemplate from the given font data used as the source of the @font-face rule. If values like weight, stretch, and style are not specified they are initialized based on the contents of the font itself.

FontTemplateCacheKey 🔒

LowercaseFontFamilyName

ResolvedGenericFontFamilies 🔒

SystemFontService

The system font service. There is one of these for every Servo instance. This is a thread, responsible for reading the list of system fonts, handling requests to match against them, and ensuring that only one copy of system font data is loaded at a time.

SystemFontServiceProxy

The public interface to the SystemFontService, used by per-Document FontContext instances.

SystemFontServiceProxySender

Enums

ComputedFontStyleDescriptor

A version of FontStyle from Stylo that is serializable. Normally this is not because the specified version of FontStyle contains floats.

FontIdentifier

SystemFontServiceMessage

Commands that the FontContext sends to the SystemFontService.