Crate fonts

Re-exports

pub use platform::LocalFontIdentifier;

pub use crate::platform::font_list::fallback_font_families;

Modules

font 🔒

font_context 🔒

font_store 🔒

font_template 🔒

glyph 🔒

platform

shaper 🔒

system_font_service 🔒

Macros

ot_tag

Structs

ByteIndex

An index that refers to a byte offset in a text run. This could the middle of a glyph.

CSSFontFaceDescriptors

This data structure represents the various optional descriptors that can be applied to a @font-face rule in CSS. These are used to create a FontTemplate from the given font data used as the source of the @font-face rule. If values like weight, stretch, and style are not specified they are initialized based on the contents of the font itself.

FallbackFontSelectionOptions

Font

FontBaseline

FontContext

The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the system font service where required.

FontData

A data structure to store data for fonts. Data is stored internally in an IpcSharedMemory handle, so that it can be send without serialization across IPC channels.

FontDescriptor

FontDescriptor describes the parameters of a Font. It represents rendering a given font template at a particular size, with a particular font-variant-caps applied, etc. This contrasts with FontTemplateDescriptor in that the latter represents only the parameters inherent in the font data (weight, stretch, etc.).

FontFamilyDescriptor

The font family parameters for font selection.

FontGroup

A FontGroup is a prioritised list of fonts for a given set of font styles. It is used by TextRun to decide which font to render a character with. If none of the fonts listed in the styles are suitable, a fallback font may be used.

FontMetrics

FontStore

FontTemplate

This describes all the information needed to create font instance handles. It contains a unique FontTemplateData structure that is platform specific.

FontTemplateDescriptor

Describes how to select a font from a given family. This is very basic at the moment and needs to be expanded or refactored when we support more of the font styling parameters.

FontTemplates

A list of font templates that make up a given font family.

GlyphData

GlyphEntry

GlyphEntry is a port of Gecko’s CompressedGlyph scheme for storing glyph data compactly.

GlyphRun

A single series of glyphs within a text run.

GlyphStore

Stores the glyph data belonging to a text run.

LowercaseFontFamilyName

RunMetrics

ShapedGlyphData

ShapedGlyphEntry

Shaper

ShapingFlags

ShapingOptions

Various options that control text shaping.

SystemFontService

The system font service. There is one of these for every Servo instance. This is a thread, responsible for reading the list of system fonts, handling requests to match against them, and ensuring that only one copy of system font data is loaded at a time.

SystemFontServiceProxy

The public interface to the SystemFontService, used by per-Document FontContext instances.

SystemFontServiceProxySender

Enums

ComputedFontStyleDescriptor

A version of FontStyle from Stylo that is serializable. Normally this is not because the specified version of FontStyle contains floats.

EmojiPresentationPreference

Whether or not font fallback selection prefers the emoji or text representation of a character. If None then either presentation is acceptable.

FontIdentifier

FontSearchScope

The scope within which we will look for a font.

GlyphInfo

SystemFontServiceMessage

Commands that the FontContext sends to the SystemFontService.

Constants

BASE

CBDT

COLR

GPOS

GSUB

KERN

LAST_RESORT_GLYPH_ADVANCE

SBIX

Traits

FontContextWebFontMethods

FontTableMethods

FontTemplateRefMethods

PlatformFontMethods

Functions

fixed_to_float 🔒

float_to_fixed 🔒

get_and_reset_text_shaping_performance_counter

Get the number of nanoseconds spent shaping text across all threads.

unicode_to_hb_script

Type Aliases

FontRef

FontTableTag

FontTemplateRef

A reference to a FontTemplate with shared ownership and mutability.

FractionalPixel

GlyphId

The id of a particular glyph within a font