fonts

Module font_context

Structs

FontCacheKey 🔒

FontContext

The FontContext represents the per-thread/thread state necessary for working with fonts. It is the public API used by the layout and paint code. It talks directly to the system font service where required.

FontFaceRuleInitiator 🔒

FontGroupCacheKey 🔒

FontParameters 🔒

RemoteWebFontDownloader 🔒

WebFontDocumentContext

Document-specific data required to fetch a web font.

WebFontDownloadState 🔒

Enums

DownloaderResponseResult 🔒

WebFontLoadInitiator 🔒

Statics

SMALL_CAPS_SCALE_FACTOR 🔒

Traits

CspViolationHandler

A callback that will be invoked on the Fetch thread if a web font download results in CSP violations. This handler will be cloned each time a new web font download is initiated.

FontContextWebFontMethods

NetworkTimingHandler

A callback that will be invoked on the Fetch thread when a web font download succeeds, providing timing information about the request.

Type Aliases

FontGroupRef

ScriptWebFontLoadFinishedCallback 🔒