Module fonts::glyph

Structs

ByteIndex

An index that refers to a byte offset in a text run. This could the middle of a glyph.

DetailedGlyph 🔒

DetailedGlyphRecord 🔒

DetailedGlyphStore 🔒

GlyphData

GlyphEntry

GlyphEntry is a port of Gecko’s CompressedGlyph scheme for storing glyph data compactly.

GlyphRun

A single series of glyphs within a text run.

GlyphStore

Stores the glyph data belonging to a text run.

Enums

GlyphInfo

Constants

FLAG_CHAR_IS_WORD_SEPARATOR 🔒

FLAG_IS_SIMPLE_GLYPH 🔒

GLYPH_ADVANCE_MASK 🔒

GLYPH_ADVANCE_SHIFT 🔒

GLYPH_COUNT_MASK 🔒

GLYPH_ID_MASK 🔒

Functions

is_simple_advance 🔒

is_simple_glyph_id 🔒

Type Aliases

GlyphId

The id of a particular glyph within a font