Module fonts::platform::freetype::library_handle

Structs

FreeTypeLibraryHandle 🔒

A FreeType library handle to be used for creating and dropping FreeType font faces. It is very important that this handle lives as long as the faces themselves, which is why only one of these is created for the entire execution of Servo and never dropped during execution.

Statics

FREETYPE_LIBRARY_HANDLE 🔒

FREETYPE_MEMORY_USAGE 🔒

Functions

ft_alloc 🔒

ft_free 🔒

ft_realloc 🔒