fonts

Trait NetworkTimingHandler

pub trait NetworkTimingHandler: Send + Debug {
    // Required methods
    fn submit_timing(&self, url: ServoUrl, response: ResourceFetchTiming);
    fn clone(&self) -> Box<dyn NetworkTimingHandler>;
}

A callback that will be invoked on the Fetch thread when a web font download succeeds, providing timing information about the request.

Required Methods

fn submit_timing(&self, url: ServoUrl, response: ResourceFetchTiming)

fn clone(&self) -> Box<dyn NetworkTimingHandler>